Refactor ModelGrid + Integrate SearchPage
