+++ Frontend/src/components/Grids/ModelsGrid/modelsGrid.module.css
@@ -1,0 +1,4 @@
+.cardsContainer {
+    margin-block: 1%;
+    padding: 1%;
+  }
